voice_bot/requirements.txt
streamlit
groq
speechrecognition
pyttsx3
python-dotenv
langchain
langchain_community
langchain-core
langchain-groq
pyaudio